[DROOLS-6870] Fix as per PR requests
